requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
cryptography==39.0.2
debugpy==1.6.6
decorator==5.1.1
Django==4.1.6
django-cors-headers==3.14.0
django-environ==0.9.0
django-rest==0.8.7
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dynaconf==3.1.12
et-xmlfile==1.1.0
executing==1.2.0
gunicorn==20.1.0
hvac==1.1.0
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
loguru==0.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyhcl==0.4.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
requests==2.28.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.9.0
urllib3==1.26.14
validators==0.20.0
wcwidth==0.2.6